project detail page - writer info component
